Improve nsx attachment handling - duplicate files
Within a single notebook now compare an attachment md5 and file name to those already written to disk and if they are already used the attachment is an exact match so do not create a duplicate attachment.  Links in note content will be to the fist file.
This avoids duplicating attachment files where the same file has been uploaded to note-station more than once within a single notebook.