Fix for bug in note station where attachments set to null.
Note station exports attachments as a null value for a notebook that was created in desktop app with notes that had encryption on and then exported from the web app.

'nul' is now detected and a message is logged as a WARNING

This may also occur with an un-encrypted notes but the flakeyness of note station syncing is preventing any testing at the moment.