Replace directory and file name cleaning.
Refactored and changed some parts of the cleaning.  Now replacing dots with a dash.  Changed to two separate approaches for file and directory as django one did not handle complex strings well, especially those that may come out of note-station or other notes.

Changed to allow unicode for all uses of the cleaning functions in yanom.

	modified:   src/config.py
	modified:   src/conversion_settings.py
	modified:   src/helper_functions.py
	modified:   src/nsx_file_converter.py
	modified:   src/sn_attachment.py
	modified:   src/sn_note_page.py
	modified:   src/sn_notebook.py
	modified:   test/test_conversion_settings.py
	modified:   test/test_helper_functoins.py
	modified:   test/test_nsx_file_converter.py
	modified:   test/test_sn_notebook.py