Handle missing json data in nsx file for values used in nsx_file_converter.py  Some tidying up of tests.
If the key or the value for the key are missing/empty for note book id's and page id's the issue is logged and nsx file is skipped by returning early.

It is highly unlikely these two keys will be missing but you never know what note station may choose to not provide.

Removed unused imports and unused values from tests.