Improve handling missing content and attachment keys in note json data.
No functional change as if no key still continues with empty string or empty dict but now with more logging of what is going on.
Move warning for attachments set to null. Was logged when creating attachments but now needs to be in the new get_json_attachment_data() method

	modified:   src/sn_note_page.py
	modified:   test/test_sn_note_page.py

	modified:   src/sn_note_page.py
	modified:   test/test_sn_note_page.py