Handle potential missing note data when adding note pages from nsx file.
Missing note data skips adding note page to list of note pages to be processed.
At end of adding pages a message is logged and printed to screen if not silent mode informing user of how many oages were not found in the nsx file by comparing id's vs the final number of pages to be processed.