Add - file cleaning options to allow user to customise name cleaning for files and directories
Add settings for allowing spaces, unicode and non-alphanumeric and uppercase characters in file names.
Also allow user to choose character(s) to replace spaces with.
Allow usr to set a maximum file or directory length. On Windows in the interactive command line a message will state if long file names are enabled or not, if not max length is restricted to 64, else 255.  255 also iused for other systems.
Settings are implemented in the name cleaning functions,  some small changes were alos implemented to improve the cleaning logic.