Add - For nsx files if images/file attachments do not have an extension add an extension where possible.
If nsx files are not syncing it has been seen that file attachments files and images may not be stored in the nsx file with a valid file name and extension.  The file wil be given a random 6 character name and wil be read to see if the file header can be used for the file type.  If it can then for recognised file types an extension is added and the link in the notes will use this.  If not recognised the file wil just be a 6 character name.

Makes use of the filetype pypi package for finding the extension.

Some tidy-up of test folder