Add - Moving exported files for html and markdown conversion part 2
Working on attachment moving.

Add questions for source and export folder paths to interactive command line so that source and export folders are always asked for, any value in the ini or command line args will be displayed as the default.

Add attachment handling to HTML to MD and MD to HTML converters.  Attachments now copied ot export folder.

Add unquote for %20 style encoding of uri's in content.

Create sets of links to use for moving files.

Simplified how note links are updated in html content in the method update_note_links to not use regex anymore, instead using urllib.parse.urlparse