Change path to logs and config.ini
Now the package builds are all using single file the yanom folder is not full of package files.  Now the config.ini and logs can be placed in the same folder as the yanom executable easier to find and more logical.

The data folder is now for data - both source and export.

Updated build scripts and docker files to match new folder for config.ini