Add - Moving exported files for html and markdown conversion part 6
Windows.... Change the sets of paths used to track attachments from using Path objects to string.  In windows the WindowsPath object changes the string representation of the path so could no longer use it to make content replacements as strings  no longer matched.  Now strings match and the strings are converted to paths as needed. This is a better way and should have been the first choice :-) The one exception is for the set copyable_absolute which is only used to move files and remains a set of path objects.