Fix - markdown to markdown conversion hyperlink may stop working
If a hyperlink is in markdown in the format <https://github.com/kevindurston21/YANOM-Note-O-Matic>  it was not maintained when converting to another Markdown format.

This error was caused by the change in image handling to use beautiful soup, which breaks the link in this 'odd' format.

Added post-processing to check and replace these links with std markdown format.

Also moved image formatting for obsidian from post to pre-processing for html and nsx conversions to minimise impact on exported markdown.

Note - These odd links are only created when the 'a' tag has the display text exactly the same as the hyperlink like this
<a href="http://github.com/kevindurston21">http://github.com/kevindurston21</a>