Fix - Obsidian to other markdown image tag paths with parenthesis in name fail
When converting from an obsidian Markdown format if the path had ( or ) in it the parsing would fail.

Re-written and now works.  Also, now includes handling of height as a value in the alt text box ![alt text|200x300] style

NOTE height is not supported in other methods/functions yet